refactor: removed the preAuthorizedCode from the issuer and moved the create deeplink function to a separate class
